# Add an optional `Volatility` argument to `NodeRule#new{Getter|Setter}`.