By default, `Volatility.NON_VOLATILE` is used.

Tested by manually inspecting the generated sources.

#446