@ben-manes 2020-08-19
Running the [GetPutBenchmark](https://github.com/ben-manes/caffeine/blob/master/caffeine/src/jmh/java/com/github/benmanes/caffeine/cache/GetPutBenchmark.java) and it appears that while reads are unaffected (as volatile is mostly free), writes do degrade. That's hard to pinpoint if noise (JIT, background work) as it maximizes the cpu either way, but over a few runs it does seem a bit slower. That's still an acceptable performance, but an fyi. I changed to run with `expireAfterWrite` and setting the `ticker` to a constant value to avoid nanoTime overheads.

The argument for making the value a volatile is weaker than the expiration timestamps (a reference cannot see any tearing). The existing logic mirrors Guava's [Suppliers.memoize](https://github.com/google/guava/blob/master/guava/src/com/google/common/base/Suppliers.java#L112-L117) which piggybacks on the fence from a preceding volatile read and publishing on the unlock. That code, as well as optimizations throughout j.u.c., were written by Martin Buchholz who should still be in your Java core team at Google for reference. While I understand the spec obsession when running a sanitizer, one does not still write C towards the PDP-11. Allowing benign races which can do no harm is common in concurrency code, especially when optimizing against memory fences and HotSpot directly.

If you have a clean environment to benchmark on (rather than my macbook!), I think that would make me more comfortable merging this in. I think any degradation should be acceptable, modest, and only in a synthetic stress test so as to have no real world implications. That, or Martin telling me to just let it though, would avoid my unease as otherwise I feel that I'd do so merely to satisfy a pedantic sanitizer.