@jamesdbowman 2020-08-19
### Summary

A benchmark on six-core workstation showed the same slowdown in writes/sec.

I am definitely not well-versed in JVM nor Caffeine internals to be able to make a convincing argument in one direction or the other; I am merely a happy consumer of this library with a thread sanitization violation filed against my team. I feel inclined to agree that the change only satisfies a pedantic sanitizer for measurable performance loss and doesn't seem ideal. Though, again, I don't know enough about JVM internals to make such an argument!

To paraphrase Martin Buchholz on the Java team: he seems to agree that using non-volatile `Unsafe` methods to access volatile fields ("piggy-backing" as it has been referred to) is a standard access pattern, and that this change is unnecessary. I can't find a GitHub account for Martin, or I'd tag him here.

@caoman, any additional thoughts on this change?

### Benchmark Environment

```
$ lscpu
Architecture:                    x86_64
CPU op-mode(s):                  32-bit, 64-bit
Byte Order:                      Little Endian
Address sizes:                   46 bits physical, 48 bits virtual
CPU(s):                          12
On-line CPU(s) list:             0-11
Thread(s) per core:              2
Core(s) per socket:              6
Socket(s):                       1
NUMA node(s):                    1
Vendor ID:                       GenuineIntel
CPU family:                      6
Model:                           85
Model name:                      Intel(R) Xeon(R) W-2135 CPU @ 3.70GHz

```

### Benchmark Results

Without change:

```
# Threads: 8 threads (1 group; 6x "readwrite_get", 2x "readwrite_put" in each group), will synchronize iterations
# Benchmark mode: Throughput, ops/time
# Benchmark: com.github.benmanes.caffeine.cache.GetPutBenchmark.readwrite
# Parameters: (cacheType = Caffeine)
```
```
Result "com.github.benmanes.caffeine.cache.GetPutBenchmark.readwrite":
  153748914.778 ±(99.9%) 19488508.653 ops/s [Average]
  (min, avg, max) = (133638324.057, 153748914.778, 178772120.076), stdev = 12890441.822
  CI (99.9%): [134260406.125, 173237423.430] (assumes normal distribution)

Secondary result "com.github.benmanes.caffeine.cache.GetPutBenchmark.readwrite:readwrite_get":
  135861653.970 ±(99.9%) 20283023.196 ops/s [Average]
  (min, avg, max) = (115874953.723, 135861653.970, 162774214.374), stdev = 13415963.999
  CI (99.9%): [115578630.774, 156144677.166] (assumes normal distribution)

Secondary result "com.github.benmanes.caffeine.cache.GetPutBenchmark.readwrite:readwrite_put":
  17887260.808 ±(99.9%) 1421922.979 ops/s [Average]
  (min, avg, max) = (15997905.702, 17887260.808, 18992842.465), stdev = 940514.011
  CI (99.9%): [16465337.828, 19309183.787] (assumes normal distribution)
```

---

With change:

```
# Threads: 8 threads (1 group; 6x "readwrite_get", 2x "readwrite_put" in each group), will synchronize iterations
# Benchmark mode: Throughput, ops/time
# Benchmark: com.github.benmanes.caffeine.cache.GetPutBenchmark.readwrite
# Parameters: (cacheType = Caffeine)
```
```
Result "com.github.benmanes.caffeine.cache.GetPutBenchmark.readwrite":
  160470832.933 ±(99.9%) 11953518.502 ops/s [Average]
  (min, avg, max) = (141874473.204, 160470832.933, 174325726.558), stdev = 7906512.374
  CI (99.9%): [148517314.431, 172424351.435] (assumes normal distribution)

Secondary result "com.github.benmanes.caffeine.cache.GetPutBenchmark.readwrite:readwrite_get":
  143351054.698 ±(99.9%) 12311279.297 ops/s [Average]
  (min, avg, max) = (124467659.876, 143351054.698, 158301622.831), stdev = 8143148.988
  CI (99.9%): [131039775.402, 155662333.995] (assumes normal distribution)

Secondary result "com.github.benmanes.caffeine.cache.GetPutBenchmark.readwrite:readwrite_put":
  17119778.235 ±(99.9%) 970339.869 ops/s [Average]
  (min, avg, max) = (15984815.749, 17119778.235, 17809465.075), stdev = 641819.743
  CI (99.9%): [16149438.366, 18090118.103] (assumes normal distribution)
```