@caoman 2020-08-19
I understand the concern. Let's do more discussion internally (mainly to get Martin's response), then follow up here.