@CLAassistant 2020-10-11
[![CLA assistant check](https://cla-assistant.io/pull/badge/not_signed)](https://cla-assistant.io/ben-manes/caffeine?pullRequest=449) <br/>Thank you for your submission! We really appreciate it. Like many open source projects, we ask that you sign our [Contributor License Agreement](https://cla-assistant.io/ben-manes/caffeine?pullRequest=449) before we can accept your contribution.<br/><sub>You have signed the CLA already but the status is still pending? Let us [recheck](https://cla-assistant.io/check/ben-manes/caffeine?pullRequest=449) it.</sub>