@ben-manes 2021-01-04
Closing due to lack of feedback. This would be done slightly differently in the v3 branch where we have replaced Unsafe with VarHandles. Thanks for the code and discussions, though.

@caoman please reopen if you have feedback and would like to revisit.